+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 
 - Mathematica programs for ordinary sums of powers
 - Mathematica documentation
+- Update Mathematica references
 
 ## [1.0.1] - 01-Jan-2026