+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Mathematica for backward differences in Eulerian numbers (Worpitzky identity)
 - Formula for backward differences in Stirling numbers
 - Mathematica for backward differences in Stirling numbers
+- Update Mathematica documentation
 
 ## [1.0.1] - 01-Jan-2026